Fix bugs in box_counting_dim and estimate_epsilon_range
box_counting_dim was also skipping all scales when any of the scales
turned out to be non-connected.

box_counting_dim was using the log(1/epsilon) rather than log(epsilon)
for the coordinates when calculating the slope in the log-log graph.

estimate_epsilon_range was using the smallest minimum distance rather
than the largest minimum distance as the distance at which two points would
always have a neighbor.

And it was returning max_neighbor_distance/2 - which is almost OK if epsilon was the side of a box and we were checking connectivity (it should be max_neighbor_distance/sqrt(2) in that case), but because we're using the direct distance between neighbors to check connectivity, this number is too small.